+++ code/gamespy/gutil.c
@@ -59,7 +59,7 @@ void gs_encode(uchar* ins, int size, uchar* result)
 
 void gs_encrypt(uchar* key, int key_len, uchar* buffer_ptr, int buffer_len)
 {
-	short counter;
+	int counter;
 	uchar x, y, xorIndex;
 	uchar state[256];
